+++ apps/meteor/client/hooks/useFormatDate.ts
@@ -4,6 +4,6 @@ import { useCallback } from 'react';
 import { formatDate } from '../lib/utils/dateFormat';
 
 export const useFormatDate = () => {
-	const format = useSetting('Message_DateFormat');
-	return useCallback((time: string | Date | number) => formatDate(time, String(format)), [format]);
+	const format = useSetting('Message_DateFormat', 'LL');
+	return useCallback((time: string | Date | number) => formatDate(time, format), [format]);
 };